requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
requests==2.31.0
pydantic==2.6.0
python-dotenv==1.0.1
redis==5.0.1
python-dateutil==2.8.2
